chore(core): move close vault screen v2
