+++ mobile-app/app/screens/AppNavigator/screens/Loans/screens/CloseVaultScreen.tsx
@@ -1,32 +1,34 @@
-import { NumberRow } from "@components/NumberRow";
-import { SubmitButtonGroup } from "@components/SubmitButtonGroup";
-import {
-  ThemedScrollView,
-  ThemedSectionTitle,
-  ThemedText,
-  ThemedView,
-} from "@components/themed";
-import { EnvironmentNetwork } from "@environment";
+import { Dispatch, useEffect, useState } from "react";
+import { Image } from "react-native";
+import { useSelector } from "react-redux";
 import { StackScreenProps } from "@react-navigation/stack";
-import { useNetworkContext } from "@shared-contexts/NetworkContext";
+import { EnvironmentNetwork } from "@environment";
 import { tailwind } from "@tailwind";
 import { translate } from "@translations";
-import { Dispatch, useEffect, useState } from "react";
+import { RootState } from "@store";
 import { hasTxQueued, transactionQueue } from "@store/transaction_queue";
 import { hasTxQueued as hasBroadcastQueued } from "@store/ocean";
-import { useSelector } from "react-redux";
-import { RootState } from "@store";
+import { onTransactionBroadcast } from "@api/transaction/transaction_commands";
+import { WhaleWalletAccount } from "@defichain/whale-api-wallet";
+import { CTransactionSegWit } from "@defichain/jellyfish-transaction/dist";
+import { useNetworkContext } from "@shared-contexts/NetworkContext";
+import { useWalletContext } from "@shared-contexts/WalletContext";
 import {
   NativeLoggingProps,
   useLogger,
 } from "@shared-contexts/NativeLoggingProvider";
-import { onTransactionBroadcast } from "@api/transaction/transaction_commands";
-import { WhaleWalletAccount } from "@defichain/whale-api-wallet";
-import { CTransactionSegWit } from "@defichain/jellyfish-transaction/dist";
-import { InfoText } from "@components/InfoText";
-import { View } from "@components";
-import { WalletAddressRow } from "@components/WalletAddressRow";
 import { useAppDispatch } from "@hooks/useAppDispatch";
+import { useAddressLabel } from "@hooks/useAddressLabel";
+import { RandomAvatar } from "@screens/AppNavigator/screens/Portfolio/components/RandomAvatar";
+import { View } from "@components";
+import {
+  ThemedScrollViewV2,
+  ThemedTextV2,
+  ThemedViewV2,
+} from "@components/themed";
+import { NumberRowV2 } from "@components/NumberRowV2";
+import { SubmitButtonGroupV2 } from "@components/SubmitButtonGroupV2";
+import CloseVaultImg from "@assets/images/loans/close_vault.png";
 import { LoanParamList } from "../LoansNavigator";
 
 type Props = StackScreenProps<LoanParamList, "CloseVaultScreen">;
@@ -43,15 +45,8 @@ export function CloseVaultScreen({ route, navigation }: Props): JSX.Element {
   const logger = useLogger();
   const [isOnPage, setIsOnPage] = useState<boolean>(true);
 
-  function onCancel(): void {
-    navigation.navigate({
-      name: "VaultDetailScreen",
-      params: {
-        vaultId: vaultId,
-      },
-      merge: true,
-    });
-  }
+  const { address } = useWalletContext();
+  const addressLabel = useAddressLabel(address);
 
   async function onSubmit(): Promise<void> {
     if (hasPendingJob || hasPendingBroadcastJob) {
@@ -75,89 +70,174 @@ export function CloseVaultScreen({ route, navigation }: Props): JSX.Element {
     };
   }, []);
 
+  const truncatedVaultId = `${vaultId.substring(0, 3)}...${vaultId.substring(
+    vaultId.length - 3,
+    vaultId.length
+  )}`;
+
   return (
-    <ThemedScrollView>
-      <SummaryHeader vaultId={vaultId} />
-      <SummaryDetails />
-      <View style={tailwind("my-4 mx-4")}>
-        <InfoText
-          testID="close_vault_info_text"
-          text={translate(
-            "screens/CloseVaultScreen",
-            "Upon closing, this Vault ID will not be usable anymore. Make sure you don’t need anything from this vault before closing."
+    <ThemedScrollViewV2
+      contentContainerStyle={tailwind("mt-16 h-full justify-between")}
+      style={tailwind("px-5")}
+    >
+      <View style={tailwind("items-center")}>
+        <Image
+          source={CloseVaultImg}
+          style={[
+            tailwind("items-center justify-center"),
+            { width: 122, height: 95 },
+          ]}
+          resizeMode="contain"
+        />
+        <ThemedTextV2
+          style={tailwind(
+            "mb-1 font-normal-v2 text-base mt-6 mx-7 text-center"
           )}
+        >
+          {translate(
+            "screens/CloseVaultScreen",
+            "Are you sure you want to close your vault {{truncatedVaultId}}?",
+            {
+              truncatedVaultId,
+            }
+          )}
+        </ThemedTextV2>
+        <SummaryDetails address={address} addressLabel={addressLabel} />
+      </View>
+
+      <View style={tailwind("mb-28 w-full items-center")}>
+        <ThemedTextV2
+          testID="transaction_details_info_text"
+          light={tailwind("text-mono-light-v2-500")}
+          dark={tailwind("text-mono-dark-v2-500")}
+          style={tailwind("mt-5 text-xs text-center font-normal-v2 mx-16")}
+        >
+          {translate(
+            "screens/CloseVaultScreen",
+            "All remaining collaterals will be returned to your wallet."
+          )}
+        </ThemedTextV2>
+        <SubmitButtonGroupV2
+          isDisabled={hasPendingJob || hasPendingBroadcastJob}
+          label={translate("screens/CloseVaultScreen", "Close vault")}
+          displayCancelBtn={false}
+          onSubmit={onSubmit}
+          buttonStyle="mt-5 mx-7"
+          title="close_vault"
         />
       </View>
-      <SubmitButtonGroup
-        isDisabled={hasPendingJob || hasPendingBroadcastJob}
-        label={translate("screens/CloseVaultScreen", "CONFIRM CLOSE VAULT")}
-        isProcessing={hasPendingJob || hasPendingBroadcastJob}
-        processingLabel={translate("screens/CloseVaultScreen", "CLOSING")}
-        onCancel={onCancel}
-        onSubmit={onSubmit}
-        displayCancelBtn
-        title="create_vault"
-      />
-    </ThemedScrollView>
+    </ThemedScrollViewV2>
   );
 }
 
-function SummaryHeader(props: { vaultId: string }): JSX.Element {
-  return (
-    <ThemedView
-      light={tailwind("bg-white border-b border-gray-300")}
-      dark={tailwind("bg-gray-800 border-b border-gray-700")}
-      style={tailwind("flex-col px-4 py-6")}
-    >
-      <ThemedText
-        light={tailwind("text-gray-500")}
-        dark={tailwind("text-gray-400")}
-        style={tailwind("mb-1 text-sm")}
-      >
-        {translate("screens/CloseVaultScreen", "You are closing Vault ID")}
-      </ThemedText>
-      <ThemedText style={tailwind("text-sm font-medium mb-1")}>
-        {props.vaultId}
-      </ThemedText>
-    </ThemedView>
-  );
-}
-
-function SummaryDetails(): JSX.Element {
+function SummaryDetails(props: {
+  addressLabel: string | null;
+  address: string | null;
+}): JSX.Element {
   const { network } = useNetworkContext();
+  const rowStyle = {
+    lhsThemedProps: {
+      light: tailwind("text-mono-light-v2-500"),
+      dark: tailwind("text-mono-dark-v2-500"),
+    },
+    rhsThemedProps: {
+      light: tailwind("text-mono-light-v2-900"),
+      dark: tailwind("text-mono-dark-v2-900"),
+    },
+  };
   return (
-    <>
-      <ThemedSectionTitle
-        text={translate("screens/CloseVaultScreen", "TRANSACTION DETAILS")}
-      />
-      <WalletAddressRow />
-      <NumberRow
-        lhs={translate("screens/CloseVaultScreen", "Fees to return")}
+    <ThemedViewV2
+      style={tailwind("p-5 border-0.5 mt-6 w-full rounded-lg-v2")}
+      dark={tailwind("border-mono-dark-v2-300")}
+      light={tailwind("border-mono-light-v2-300")}
+    >
+      <View style={tailwind("flex flex-row pb-5 items-center justify-between")}>
+        <ThemedTextV2>
+          <ThemedTextV2
+            style={tailwind("text-sm font-normal-v2")}
+            light={rowStyle.lhsThemedProps.light}
+            dark={rowStyle.lhsThemedProps.dark}
+          >
+            {translate("screens/CloseVaultScreen", "Wallet")}
+          </ThemedTextV2>
+        </ThemedTextV2>
+        <ThemedViewV2
+          dark={tailwind("bg-mono-dark-v2-200")}
+          light={tailwind("bg-mono-light-v2-200")}
+          style={tailwind(
+            "rounded-full pl-1 pr-2.5 py-1 flex flex-row items-center overflow-hidden ml-2"
+          )}
+        >
+          <RandomAvatar name={props.address} size={20} />
+          <ThemedTextV2
+            ellipsizeMode="middle"
+            numberOfLines={1}
+            style={[
+              tailwind("text-sm font-normal-v2 ml-1", {
+                "mx-2": props.addressLabel !== undefined,
+                "text-right w-28": props.addressLabel === undefined,
+              }),
+              {
+                minWidth: 10,
+                maxWidth: props.addressLabel !== undefined ? 101 : 0,
+              },
+            ]}
+            testID="wallet_address"
+          >
+            {props.addressLabel ?? props.address}
+          </ThemedTextV2>
+        </ThemedViewV2>
+      </View>
+      <NumberRowV2
+        containerStyle={{
+          style: tailwind("pb-5 flex-row items-start w-full bg-transparent"),
+        }}
+        lhs={{
+          value: translate("screens/CloseVaultScreen", "Fees to return"),
+          testID: "fees_to_return_text_lhs",
+          themedProps: {
+            light: rowStyle.lhsThemedProps.light,
+            dark: rowStyle.lhsThemedProps.dark,
+          },
+        }}
         rhs={{
           value:
             network === EnvironmentNetwork.MainNet ||
             network === EnvironmentNetwork.TestNet
               ? 1
               : 0.5,
-          testID: "fees_to_return_text",
-          suffixType: "text",
-          suffix: "DFI",
+          testID: "fees_to_return_text_rhs",
+          suffix: " DFI",
+          themedProps: {
+            light: rowStyle.rhsThemedProps.light,
+            dark: rowStyle.rhsThemedProps.dark,
+          },
         }}
       />
-      <NumberRow
-        lhs={translate("screens/CloseVaultScreen", "Fees to burn")}
+      <NumberRowV2
+        lhs={{
+          value: translate("screens/CloseVaultScreen", "Fees to burn"),
+          testID: "fees_to_burn_text_lhs",
+          themedProps: {
+            light: rowStyle.lhsThemedProps.light,
+            dark: rowStyle.lhsThemedProps.dark,
+          },
+        }}
         rhs={{
           value:
             network === EnvironmentNetwork.MainNet ||
             network === EnvironmentNetwork.TestNet
               ? 1
               : 0.5,
-          testID: "fees_to_burn_text",
-          suffixType: "text",
-          suffix: "DFI",
+          testID: "fees_to_burn_text_rhs",
+          suffix: " DFI",
+          themedProps: {
+            light: rowStyle.rhsThemedProps.light,
+            dark: rowStyle.rhsThemedProps.dark,
+          },
         }}
       />
-    </>
+    </ThemedViewV2>
   );
 }
 
@@ -184,15 +264,19 @@ async function closeVault(
       return new CTransactionSegWit(signed);
     };
 
+    const truncatedVaultId = `${vaultId.substring(0, 3)}...${vaultId.substring(
+      vaultId.length - 3,
+      vaultId.length
+    )}`;
+
     dispatch(
       transactionQueue.actions.push({
         sign: signer,
-        title: translate("screens/CloseVaultScreen", "Closing vault"),
-        description: translate(
+        title: translate(
           "screens/CloseVaultScreen",
-          "You are about to close vault {{vaultId}}",
+          "Closing vault {{vaultId}}",
           {
-            vaultId,
+            vaultId: truncatedVaultId,
           }
         ),
         drawerMessages: {
@@ -200,10 +284,17 @@ async function closeVault(
             "screens/OceanInterface",
             "Preparing to close vault…"
           ),
-          waiting: translate("screens/OceanInterface", "Closing vault…"),
+          waiting: translate(
+            "screens/CloseVaultScreen",
+            "Closing vault {{vaultId}}",
+            {
+              vaultId: truncatedVaultId,
+            }
+          ),
           complete: translate(
-            "screens/OceanInterface",
-            "Vault successfully closed"
+            "screens/CloseVaultScreen",
+            "Closed vault {{vaultId}}",
+            { vaultId: truncatedVaultId }
           ),
         },
         onBroadcast,
